fixed restructure and logout
