+++ app/__init__.py
@@ -1,14 +1,13 @@
 import logging
+from logging.handlers import SMTPHandler, RotatingFileHandler
+import os
 from flask import Flask, request, current_app
-from config import Config
 from flask_sqlalchemy import SQLAlchemy
 from flask_mail import Mail
 from flask_migrate import Migrate
 from flask_moment import Moment
 from flask_login import LoginManager
-from logging.handlers import SMTPHandler
-from logging.handlers import RotatingFileHandler
-import os
+from config import Config
 
 db = SQLAlchemy()
 migrate = Migrate()
@@ -30,6 +29,9 @@ def create_app(config_class=Config):
 
   from app.errors import bp as errors_bp
   app.register_blueprint(errors_bp)
+
+  from app.auth import bp as auth_bp
+  app.register_blueprint(auth_bp, url_prefix='/auth')
 
   from app.main import bp as main_bp
   app.register_blueprint(main_bp)
