+++ app/main/routes.py
@@ -58,17 +58,6 @@ def user(username):
 def about():
   return render_template('about.html', title='About')
 
-@bp.route('/delete_profile')
-def delete_profile():
-  id = request.args.get('id')
-  logout_user()
-  usersPosts = Post.query.filter(Post.user_id == id)
-  for post in usersPosts:
-    db.session.delete(post) 
-  User.query.filter(User.id == id).delete()
-  db.session.commit()
-  return redirect(url_for('main.index'))
-
 @bp.route('/edit_profile', methods=['GET', 'POST'])
 @login_required
 def edit_profile():
