+++ app/templates/base.html
@@ -30,6 +30,7 @@
     <a href="{{ url_for('auth.login') }}">login</a>
     {% else %}
     <a href="{{ url_for('main.user', username=current_user.username) }}">Profile</a>
+    <a href="{{ url_for('auth.logout') }}">logout</a>
     {% endif %}  
   </footer>
   </body>
