+++ app/templates/user.html
@@ -10,7 +10,7 @@
         {% if user.last_seen %}<p>Last seen on: {{ moment(user.last_seen).format('LLL') }}</p>{% endif %}
         {% if user == current_user %}
         <p><a href="{{ url_for('main.edit_profile') }}">Edit your profile</a></p>
-        <p><a href="{{ url_for('main.delete_profile', id=user.id) }}">Delete your profile</a></p>
+        <p><a href="{{ url_for('auth.delete_profile', id=user.id) }}">Delete your profile</a></p>
         {% endif %}
       </td>
     </tr>
